requirements.txt
instaloader
requests
Pillow
torch
transformers
openai
python-dotenv
instagrapi
tqdm
diffusers 
transformers